requirements.txt
streamlit==1.20.0
openai-whisper==20230308
setuptools-rust==1.5.2